Add Suspense wrapper to tab pane
