+++ app/src/components/Main.tsx
@@ -5,9 +5,12 @@ import React, {
   memo,
   useState,
   useCallback,
+  Suspense,
 } from "react";
+import CurrentTab from "./CurrentTab";
 import Graph from "./Graph";
 import GraphWrapper from "./GraphWrapper";
+import Loading from "./Loading";
 import TabPane from "./TabPane";
 import TextResizer from "./TextResizer";
 
@@ -23,7 +26,11 @@ const Main = memo(
     const trigger = useCallback(() => triggerResize((n) => n + 1), []);
     return (
       <>
-        <TabPane triggerResize={trigger}>{children}</TabPane>
+        <TabPane triggerResize={trigger}>
+          <Suspense fallback={<Loading />}>
+            <CurrentTab>{children}</CurrentTab>
+          </Suspense>
+        </TabPane>
         <GraphWrapper>
           <Graph
             textToParse={textToParse}
